move the phase tracker to an API-only package
Importing internal/traces from the translators pulled the OTLP exporters into
every consumer of the Gateway API translator, including egctl and the separate
test module, which then failed to build without a new indirect requirement in
test/go.mod. Nothing about recording a span needs the exporters.

The tracker now lives in internal/traces/phase, which imports only the
OpenTelemetry API, so the dependency footprint of the instrumented packages is
unchanged and both modules stay tidy.
