+++ internal/xds/translator/translator.go
@@ -36,7 +36,7 @@ import (
 	extensionTypes "github.com/envoyproxy/gateway/internal/extension/types"
 	"github.com/envoyproxy/gateway/internal/ir"
 	"github.com/envoyproxy/gateway/internal/logging"
-	"github.com/envoyproxy/gateway/internal/traces"
+	"github.com/envoyproxy/gateway/internal/traces/phase"
 	"github.com/envoyproxy/gateway/internal/utils"
 	"github.com/envoyproxy/gateway/internal/utils/proto"
 	"github.com/envoyproxy/gateway/internal/xds/types"
@@ -174,7 +174,7 @@ func (t *Translator) Translate(ctx context.Context, xdsIR *ir.Xds) (*types.Resou
 
 	// Each phase ends its own span; the deferred call closes and marks the phase that
 	// was in flight when a phase panics, so the failing phase stays in the trace.
-	phases := traces.NewPhaseTracker(ctx, tracer)
+	phases := phase.NewTracker(ctx, tracer)
 	defer phases.EndInFlight()
 
 	t.backendIndex = newBackendClusterIndex(xdsIR)
